api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
boto3==1.36.7
Pillow==11.1.0
marshmallow==3.23.2
